requirements.txt
Django==1.9.12
MySQL-python==1.2.5
Pillow==3.4.2
django-oscar
django_compressor
django-constance
django-nested-inline
pdfkit
